+++ torchays/utils/areaUtils.py
@@ -6,6 +6,7 @@ import torch
 from scipy.optimize import minimize
 
 from torchays.modules.base import AysBaseModule
+from torchays.utils.minfunc import func, funcJac, conFunc, conFunc1, conJac, boundFun, boundJac, cfunc, jacCcfunc, ccfunc, jacCfunc
 
 
 class WapperArea(object):
@@ -57,11 +58,13 @@ class WapperArea(object):
 
 class AnalysisReLUNetUtils(object):
     """
-    AnalysisReLUNetUtils.
+    AnalysisReLUNetUtils needs to ensure that the net has the function:
+        >>> def forward_graph_Layer(*args, layer=layerNum)
+            # 'layer' is a int parameters which can get the layer weight and bias graph.
 
     args:
         device: torch.device
-            GPU or CPU to get the graph;
+            GPU or CPU to get the graph from the network;
         logger: def info(...)
             print the information (Default: print in console)(logger.info(...)).
     """
@@ -354,75 +357,3 @@ class AnalysisReLUNetUtils(object):
         """
         assert self.saveArea, "Not save some area infomation"
         return self.areaFuncs, self.areas, self.points
-
-# ================================================================
-# Minimize function.
-
-
-def func(function):
-    """  (aX+b)^2 """
-    def xx(x):
-        return np.square(np.matmul(x, function[:-1]) + function[-1])
-    return xx
-
-
-def funcJac(function):
-    """ {2a_i(aX+b);i in (0, n)} """
-    def xx(x):
-        return np.array([((np.matmul(x, function[:-1]) + function[-1]) * function[i] * 2) for i in range(function.shape[0]-1)])
-    return xx
-
-
-def conFunc(function):
-    """ aX+b """
-    def xx(x):
-        return np.matmul(x, function[:-1]) + function[-1]
-    return xx
-
-
-def conFunc1(function):
-    """ aX+b """
-    def xx(x):
-        return np.matmul(x, function[:-1]) + function[-1] - 1e-10
-    return xx
-
-
-def conJac(function):
-    """ {a_i;i in (0, n)} """
-    def xx(x):
-        return function[:-1]
-    return xx
-
-
-def boundFun(x):
-    return np.sum(np.square(x)) - 1e-32
-
-
-def boundJac(x):
-    return np.array([2*x[i] for i in range(x.shape[0])])
-
-
-def cfunc(function, normA):
-    def xx(x):
-        return np.matmul(x[:-1], function[:-1]) + function[-1] - normA * x[-1]
-    return xx
-
-
-def jacCfunc(function, normA):
-    def xx(x):
-        return np.append(function[:-1], -normA)
-    return xx
-
-
-def ccfunc(funcion):
-    def xx(x):
-        return -x[-1]
-    return xx
-
-
-def jacCcfunc(function):
-    def xx(x):
-        output = np.zeros_like(function)
-        output[-1] -= 1
-        return output
-    return xx
